src/py-mqtt-sample/requirements.txt
paho-mqtt==1.5.0
flask